Refactor CreativeTokensGrid to utilize MemoizedTokenCard for improved performance
- Replaced HottestTokenCard with MemoizedTokenCard to enhance rendering efficiency and maintain flip states across re-renders.
- Removed redundant helper functions and components, consolidating logic within MemoizedTokenCard.
- Introduced a new state management for flip states to prevent loss during updates.
- Updated token rendering logic to streamline the display of trending tokens.

Test plan:
- Verify that token flipping functionality works correctly and maintains state.
- Ensure that the MemoizedTokenCard displays tokens with the expected visual effects and data.
- Confirm that performance improvements are noticeable in the CreativeTokensGrid.
